+++ src/FakeXrmEasy.Plugins/OrganizationRequestNameConstants.cs
@@ -1,0 +1,20 @@
+namespace FakeXrmEasy.Plugins
+{
+    internal static class OrganizationRequestNameConstants
+    {
+        internal const string CREATE = "Create";
+        internal const string UPDATE = "Update";
+        internal const string DELETE = "Delete";
+        internal const string RETRIEVE = "Retrieve";
+        internal const string RETRIEVE_MULTIPLE = "RetrieveMultiple";
+        internal const string UPSERT = "Upsert";
+
+        internal const string CREATE_MULTIPLE = "CreateMultiple";
+        internal const string UPDATE_MULTIPLE = "UpdateMultiple";
+        internal const string UPSERT_MULTIPLE = "UpsertMultiple";
+        
+        internal const string SEND_EMAIL = "SendEmail";
+        internal const string SEND_TEMPLATE = "SendTemplate";
+        internal const string SEND_FAX = "SendFax";
+    }
+}