+++ tests/FakeXrmEasy.Plugins.Tests/PluginSteps/PluginStepValidatorTests/PluginStepValidatorSendTests.cs
@@ -1,0 +1,60 @@
+﻿using DataverseEntities;
+using FakeXrmEasy.Abstractions.Plugins.Enums;
+using FakeXrmEasy.Plugins.PluginSteps;
+using FakeXrmEasy.Plugins.PluginSteps.InvalidRegistrationExceptions;
+using Xunit;
+
+namespace FakeXrmEasy.Plugins.Tests.PluginSteps.PluginStepValidatorTests
+{
+    public class PluginStepValidatorSendTests
+    {
+        private readonly PluginStepValidator _validator;
+
+        public PluginStepValidatorSendTests()
+        {
+            _validator = new PluginStepValidator();
+        }
+
+        [Theory]
+        [InlineData(MessageNameConstants.Send, EntityLogicalNameConstants.Email, ProcessingStepStage.Prevalidation, ProcessingStepMode.Synchronous)]
+        [InlineData(MessageNameConstants.Send, EntityLogicalNameConstants.Fax, ProcessingStepStage.Prevalidation, ProcessingStepMode.Synchronous)]
+        [InlineData(MessageNameConstants.Send, EntityLogicalNameConstants.Template, ProcessingStepStage.Prevalidation, ProcessingStepMode.Synchronous)]
+        [InlineData(MessageNameConstants.Send, EntityLogicalNameConstants.Email, ProcessingStepStage.Preoperation, ProcessingStepMode.Synchronous)]
+        [InlineData(MessageNameConstants.Send, EntityLogicalNameConstants.Fax, ProcessingStepStage.Preoperation, ProcessingStepMode.Synchronous)]
+        [InlineData(MessageNameConstants.Send, EntityLogicalNameConstants.Template, ProcessingStepStage.Preoperation, ProcessingStepMode.Synchronous)]
+        [InlineData(MessageNameConstants.Send, EntityLogicalNameConstants.Email, ProcessingStepStage.Postoperation, ProcessingStepMode.Synchronous)]
+        [InlineData(MessageNameConstants.Send, EntityLogicalNameConstants.Fax, ProcessingStepStage.Postoperation, ProcessingStepMode.Synchronous)]
+        [InlineData(MessageNameConstants.Send, EntityLogicalNameConstants.Template, ProcessingStepStage.Postoperation, ProcessingStepMode.Synchronous)]
+        [InlineData(MessageNameConstants.Send, EntityLogicalNameConstants.Email, ProcessingStepStage.Postoperation, ProcessingStepMode.Asynchronous)]
+        [InlineData(MessageNameConstants.Send, EntityLogicalNameConstants.Fax, ProcessingStepStage.Postoperation, ProcessingStepMode.Asynchronous)]
+        [InlineData(MessageNameConstants.Send, EntityLogicalNameConstants.Template, ProcessingStepStage.Postoperation, ProcessingStepMode.Asynchronous)]
+        public void Should_return_valid_registrations_for_supported_entity_logical_names(string messageName, string entityLogicalName, ProcessingStepStage stage, ProcessingStepMode mode)
+        {
+            var pluginStepDefinition = new PluginStepDefinition()
+            {
+                MessageName = messageName,
+                EntityLogicalName = entityLogicalName,
+                Stage = stage,
+                Mode = mode
+            };
+            Assert.True(_validator.IsValid(pluginStepDefinition));
+        }
+
+        [Theory]
+        [InlineData(MessageNameConstants.Send, EntityLogicalNameConstants.Account, ProcessingStepStage.Prevalidation, ProcessingStepMode.Synchronous)]
+        [InlineData(MessageNameConstants.Send, EntityLogicalNameConstants.Account, ProcessingStepStage.Preoperation, ProcessingStepMode.Synchronous)]
+        [InlineData(MessageNameConstants.Send, EntityLogicalNameConstants.Account, ProcessingStepStage.Postoperation, ProcessingStepMode.Synchronous)]
+        [InlineData(MessageNameConstants.Send, EntityLogicalNameConstants.Account, ProcessingStepStage.Postoperation, ProcessingStepMode.Asynchronous)]
+        public void Should_return_invalid_registrations_for_unsupported_entity_logical_names(string messageName, string entityLogicalName, ProcessingStepStage stage, ProcessingStepMode mode)
+        {
+            var pluginStepDefinition = new PluginStepDefinition()
+            {
+                MessageName = messageName,
+                EntityLogicalName = entityLogicalName,
+                Stage = stage,
+                Mode = mode
+            };
+            Assert.Throws<InvalidPrimaryEntityNameException>(() => _validator.IsValid(pluginStepDefinition));
+        }
+    }
+}