+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [3.7.0]
+
+### Changed
+
+- Adds PreImage and PostImage support for Assign message - https://github.com/DynamicsValue/fake-xrm-easy/issues/187
+- Pipeline Simulation: Extended PreImage and PostImage support for these messages: Send, - https://github.com/DynamicsValue/fake-xrm-easy/issues/187
+- Pipeline Simulation: Extended support for new messages that don't match request names (Send message, and SendEmail, SendTemplate, SendFax requests) - https://github.com/DynamicsValue/fake-xrm-easy/issues/185
+
 ## [3.6.0]
 
 ### Changed